Refactor pattern and domain specific methods of BlockerService into a DomainBlockerService and a PatternBlockerService
