+++ eblocker-icapserver/src/main/java/org/eblocker/server/common/blocker/BlockerIdTypeIdCache.java
@@ -18,12 +18,14 @@ package org.eblocker.server.common.blocker;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
+import com.google.inject.Singleton;
 
 /**
  * Creates an in-memory cache for mapping tuples of {Type, ID} of blockers
  * to temporary IDs. These IDs start at 0 and are only meaningful within the
  * process.
  */
+@Singleton
 public class BlockerIdTypeIdCache {
 
     private final BiMap<TypeId, Integer> pseudoIdByTypeId = HashBiMap.create();
