+++ eblocker-icapserver/src/main/java/org/eblocker/server/common/util/HttpClient.java
@@ -44,7 +44,8 @@ public class HttpClient {
             if (response.getStatusLine().getStatusCode() != 200) {
                 throw new IOException("Could not download " + url + ". Status " + response.getStatusLine().getStatusCode() + " returned");
             }
-            return new ByteArrayInputStream(ByteStreams.toByteArray(response.getEntity().getContent()));
+            //return new ByteArrayInputStream(ByteStreams.toByteArray(response.getEntity().getContent()));
+            return response.getEntity().getContent();
         }
     }
 }
